WIP testing multiple scenes with rendering

+++ Framework/include/Gamejam/Demo/Behaviours/DemoRenderBehaviour.h
@@ -9,7 +9,7 @@ namespace jam::demo
 	class DemoRenderBehaviour final : public jecs::ISystemBehaviour<DemoRenderComponent>
 	{
 	public:
-		uint32_t m_zoom = 1;
+		float m_zoom = 1;
 		double m_angle = 0;
 		float xOffset = 0, yOffset = 0;
 		
@@ -19,7 +19,6 @@ namespace jam::demo
 
 	private:
 		SDL_Texture* m_texture;
-
 
 		void PreRender() const;
 		void PostRender() const;
